Move ring files into ring dir
