+++ lib/ring/index.js
